splitted code to files, added tests
