--- .gitignore
@@ -1,76 +1,0 @@
-# Prerequisites
-*.d
-
-# Compiled Object files
-*.slo
-*.lo
-*.o
-*.obj
-
-# Precompiled Headers
-*.gch
-*.pch
-
-# Compiled Dynamic libraries
-*.so
-*.dylib
-*.dll
-
-# Fortran module files
-*.mod
-*.smod
-
-# Compiled Static libraries
-*.lai
-*.la
-*.a
-*.lib
-
-# Executables
-*.exe
-*.out
-*.app
-libtool
-
-*.m4
-*.in
-*.pc
-*.libs
-*.deps
-*.cache
-*.dirstamp
-*.swp
-*.orig
-*.rej
-
-bin/
-out/
-
-# CMake
-cmake-build-*/
-
-config*
-Makefile
-build-aux
-test_tmp_*
-libtool
-docs/reference*
-docs/*.html
-docs/*.png
-docs/*-ref.rst
-
-.DS_Store
-.idea
-*~
-
-# Compile and link flag files
-bindings/python/*flags
-
-# Logs
-libtorrent_logs*
-*.log
-
-# Python related files and dirs
-*.pyc
-dist
-*.egg-info
