+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "libs/gtest"]
+	path = libs/gtest
+	url = https://github.com/google/googletest.git
